requirements.txt
fastapi
uvicorn[standard]
torch==2.1.2
torchvision
opencv-python-headless
pillow
numpy
ultralytics<=8.3.40
git+https://github.com/openai/CLIP.git
